setup: detect unsio lib if available in python
If the unsio python library is installed via pip, for example
in a virtualenv, the test does not find it. Probable culprit
is the `#include <uns.h>` is not found.

So before attempting to detect unsio with compilation, check
first if it is available from python via a simple import.

Signed-off-by: Vincent Legoll <vincent.legoll@iphc.cnrs.fr>